requirements.txt
pygame>=2.5.2
pillow
pystray
